Refactor package structure for Auth and Onboarding features
This commit reorganizes the Auth and Onboarding modules into a more structured presentation/domain/data architecture and updates the manifest to correctly set the application entry point.

- **Package Refactoring**:
    - Relocated Auth-related views, presenters, and contracts to `features.auth.presentation`.
    - Moved `User` and `OnboardingItem` models into their respective `domain.model` packages.
    - Reorganized Onboarding views, adapters, presenters, and contracts into `features.onboarding.presentation`.
    - Updated all imports and references in `nav_auth.xml` and Java classes to reflect the new package structure.

- **Manifest & Navigation**:
    - Updated `AndroidManifest.xml` to set `SplashActivity` as the launcher activity.
    - Changed `DashboardActivity` to `exported="false"` and removed its intent filter.
    - Updated activity names in the manifest to match the new package paths.

- **Cleanup**:
    - Removed unused imports of `MainActivity` and `User` model in several fragments and contracts.
