+++ app/src/main/AndroidManifest.xml
@@ -17,28 +17,28 @@
         android:theme="@style/Theme.RecipeAndroidProject">
         <activity
             android:name=".features.dashboard.view.DashboardActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity
-            android:name=".features.auth.view.AuthActivity"
-            android:exported="false" />
-        <activity
-            android:name=".features.onboarding.view.OnboardingActivity"
-            android:exported="false" />
-        <activity
-            android:name=".SplashActivity"
-            android:exported="true"
-            android:theme="@style/Theme.App.Splash">
+            android:exported="false">
 <!--            <intent-filter>-->
 <!--                <action android:name="android.intent.action.MAIN" />-->
 
 <!--                <category android:name="android.intent.category.LAUNCHER" />-->
 <!--            </intent-filter>-->
+        </activity>
+        <activity
+            android:name=".features.auth.presentation.view.AuthActivity"
+            android:exported="false" />
+        <activity
+            android:name=".features.onboarding.presentation.view.OnboardingActivity"
+            android:exported="false" />
+        <activity
+            android:name=".SplashActivity"
+            android:exported="true"
+            android:theme="@style/Theme.App.Splash">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
         <activity
             android:name=".MainActivity"
